rm executeUsercode
