+++ src/cmd/cafecoder-judge/Makefile
@@ -1,0 +1,2 @@
+judge: main.go
+	go build --race -o ./../../..
