--- src/executeUsercode/Makefile
@@ -1,2 +1,0 @@
-executeUsercode:executeUsercode.go
-	go build -o ./../..
