--- src/judge/cmd/cafecoder-judge/Makefile
@@ -1,2 +1,0 @@
-judge: main.go
-	go build --race -o ./../../../..
