+++ src/types/json.go
@@ -1,0 +1,16 @@
+package types
+
+type CmdResultJSON struct {
+	SessionID  string `json:"sessionID"`
+	Time       int    `json:"time"`
+	Result     bool   `json:"result"`
+	ErrMessage string `json:"errMessage"`
+	MemUsage   int    `json:"memUsage"`
+}
+
+type RequestJSON struct {
+	SessionID string `json:"sessionID"`
+	Cmd       string `json:"cmd"`
+	Mode      string `json:"mode"` //Mode ... "judge" or "other"
+	DirName   string `json:"dirName"`
+}
